apps/res-sell/shopify-graphql/requirements.txt
allure-pytest
orjson
statsd-exporter==3.2.1
jsonschema
ShopifyAPI
